Split out overflow handling into its own module
